fixed LongConverter https://github.com/w11k/lsql/issues/23

+++ lsql-core/src/main/java/com/w11k/lsql/converter/sqltypes/LongConverter.java
@@ -16,7 +16,7 @@ public class LongConverter extends Converter {
 
     @Override
     protected void setValue(LSql lSql, PreparedStatement ps, int index, Object val) throws SQLException {
-        ps.setLong(index, (Integer) val);
+        ps.setLong(index, (Long) val);
     }
 
     @Override
